+++ docs/CHANGELOG.md
@@ -5,6 +5,7 @@ Changelog
 ---------------------------
 - Enh #415: Display "Edit template" menu on structure view
 - Enh #417: Allow templates with option "Allow in spaces" also for global pages
+- Enh #419: Allow to change category of custom page
 
 1.11.0-beta.3 (June 3, 2025)
 ----------------------------